+++ packages/feathers/src/application.ts
@@ -1,7 +1,9 @@
 import version from './version'
-import { EventEmitter, stripSlashes, createDebug, HOOKS, hooks, middleware } from './dependencies'
+import { EventEmitter } from 'events'
+import { stripSlashes, createDebug } from '@feathersjs/commons'
+import { hooks, middleware } from '@feathersjs/hooks'
 import { eventHook, eventMixin } from './events'
-import { hookMixin } from './hooks/index'
+import { hookMixin } from './hooks'
 import { wrapService, getServiceOptions, protectedMethods } from './service'
 import {
   FeathersApplication,
@@ -11,28 +13,26 @@ import {
   ServiceInterface,
   Application,
   FeathersService,
-  HookMap,
   ApplicationHookOptions
 } from './declarations'
-import { enableRegularHooks } from './hooks/regular'
+import { enableHooks } from './hooks'
 
 const debug = createDebug('@feathersjs/feathers')
 
-export class Feathers<Services, Settings> extends EventEmitter implements FeathersApplication<Services, Settings> {
+export class Feathers<Services, Settings>
+  extends EventEmitter
+  implements FeathersApplication<Services, Settings>
+{
   services: Services = {} as Services
   settings: Settings = {} as Settings
   mixins: ServiceMixin<Application<Services, Settings>>[] = [hookMixin, eventMixin]
   version: string = version
   _isSetup = false
-  appHooks: HookMap<Application<Services, Settings>, any> = {
-    [HOOKS]: [eventHook as any]
-  }
 
-  private regularHooks: (this: any, allHooks: any) => any
+  protected registerHooks: (this: any, allHooks: any) => any
 
   constructor() {
     super()
-    this.regularHooks = enableRegularHooks(this)
     hooks(this, {
       setup: middleware().params('server').props({
         app: this
@@ -40,6 +40,10 @@ export class Feathers<Services, Settings> extends EventEmitter implements Feathe
       teardown: middleware().params('server').props({
         app: this
       })
+    })
+    this.registerHooks = enableHooks(this)
+    this.registerHooks({
+      around: [eventHook]
     })
   }
 
@@ -125,19 +129,16 @@ export class Feathers<Services, Settings> extends EventEmitter implements Feathe
   hooks(hookMap: ApplicationHookOptions<this>) {
     const untypedMap = hookMap as any
 
-    if (untypedMap.before || untypedMap.after || untypedMap.error) {
-      this.regularHooks(untypedMap)
+    if (untypedMap.before || untypedMap.after || untypedMap.error || untypedMap.around) {
+      // regular hooks for all service methods
+      this.registerHooks(untypedMap)
     } else if (untypedMap.setup || untypedMap.teardown) {
+      // .setup and .teardown application hooks
       hooks(this, untypedMap)
-    } else if (Array.isArray(hookMap)) {
-      this.appHooks[HOOKS].push(...(hookMap as any))
     } else {
-      const methodHookMap = hookMap as HookMap<Application<Services, Settings>, any>
-
-      Object.keys(methodHookMap).forEach((key) => {
-        const methodHooks = this.appHooks[key] || []
-
-        this.appHooks[key] = methodHooks.concat(methodHookMap[key])
+      // Other registration formats are just `around` hooks
+      this.registerHooks({
+        around: untypedMap
       })
     }