+++ packages/feathers/src/dependencies.ts
@@ -1,5 +1,0 @@
-import { EventEmitter } from 'events'
-export * from '@feathersjs/commons'
-export * from '@feathersjs/hooks'
-
-export { EventEmitter }