+++ packages/rest-client/test/declarations.ts
@@ -1,6 +1,15 @@
-import { CustomMethods } from '@feathersjs/feathers'
+import { CustomMethod } from '@feathersjs/feathers'
 import { RestService } from '../src'
 
+type Data = { message: string }
+type Result = {
+  data: Data
+  provider: string
+  type: string
+}
+
 export type ServiceTypes = {
-  todos: RestService & CustomMethods<{ customMethod: any }>
+  todos: RestService & {
+    customMethod: CustomMethod<Data, Result>
+  }
 }